Make the capture more flexible by having sent and received methods which can be overridden to support arbitrary behavior.
In RawNetCapture, added the ability to capture the exact interchange of sent and received data, mixed together.

Instead of requiring debug_output to be RawNetCapture, just check if it responds to :sent or :received methods

Moved the capturing of the full body of sent and received data into class RawHTTPCapture, since this functionality is most useful for capturing HTTP request and response.